requirements.txt
django==3.0.7
djangorestframework==3.15.0
cloudinary==1.40.0
psycopg2==2.8.5
python-dotenv==1.0.1
drf-yasg==1.21.7
pytest==8.2.0
pytest-django==4.8.0
celery==5.4.0
redis==5.0.4
